relocate schemas in a separate folder + add alias
